backend/app/utils/ai_integration.py
from typing import Dict
import requests
import os
from dotenv import load_dotenv

load_dotenv()

class AIHairStylist:
    """Integração com IA para sugestões de cabelo (Hugging Face - Gratuito)"""
    
    def __init__(self):
        self.api_key = os.getenv("HUGGINGFACE_API_KEY")
        self.api_url = "https://api-inference.huggingface.co/models/dima806/hair_segmentation"
    
    def analyze_face(self, image_path: str) -> Dict:
        """
        Analisar foto e gerar sugestões
        Nota: Esta é uma implementação básica com IA gratuita.
        Para produção, integrar com Claude API posteriormente.
        """
        try:
            # Ler imagem
            with open(image_path, "rb") as f:
                image_data = f.read()
            
            # Para demonstração, retornar sugestões baseadas em análise básica
            # Em produção, fazer chamada real para API
            suggestions = self._generate_suggestions_demo()
            
            return {
                "success": True,
                "suggestions": suggestions,
                "message": "Análise concluída com sucesso"
            }
            
        except Exception as e:
            return {
                "success": False,
                "error": str(e),
                "message": "Erro ao analisar imagem"
            }
    
    def _generate_suggestions_demo(self) -> Dict:
        """
        Gerar sugestões de demonstração
        TODO: Integrar com Claude API para sugestões reais
        """
        return {
            "cortes_sugeridos": [
                "Corte em camadas médio",
                "Bob moderno com franja lateral",
                "Corte pixie estilizado",
                "Long bob (lob) texturizado"
            ],
            "cores_sugeridas": [
                "Castanho chocolate com reflexos caramelo",
                "Loiro mel natural",
                "Ruivo acobreado",
                "Preto azulado com brilho"
            ],
            "estilos_recomendados": [
                "Ondulado natural",
                "Liso com pontas viradas",
                "Cacheado definido",
                "Semi-preso moderno"
            ],
            "cores_esmalte": [
                "Vermelho clássico",
                "Nude rosado",
                "Roxo metalizado",
                "Francesinha moderna"
            ],
            "confianca": 0.85,
            "observacoes": "Análise baseada em formato de rosto e tom de pele"
        }
    
    async def analyze_with_claude(self, image_path: str) -> Dict:
        """
        Método preparado para integração futura com Claude API
        Descomente e configure quando tiver API key do Claude
        """
        pass

# Instância global
ai_stylist = AIHairStylist()