backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
Pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy==2.1.3
requests==2.32.3
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0